reformatted the file structure. added some login signup hooks added error and success toasts. added server side routes and changed client side routes
