+++ Client/src/components/baseLayout/types.ts
@@ -1,0 +1,6 @@
+import {ReactNode} from "react";
+
+export interface BaseLayoutProps {
+    children: ReactNode,
+    admin?: boolean
+}
