+++ Client/src/index.tsx
@@ -5,6 +5,7 @@ import {ThemeProvider} from "@mui/material";
 import {lightTheme} from "./themes/themes";
 import {QueryClient, QueryClientProvider} from "react-query";
 import {SnackbarProvider} from "notistack";
+import CssBaseline from "@mui/material/CssBaseline";
 
 const root = ReactDOM.createRoot(
     document.getElementById('root') as HTMLElement
@@ -15,6 +16,7 @@ root.render(
     <ThemeProvider theme={lightTheme}>
         <QueryClientProvider client={queryClient}>
             <SnackbarProvider maxSnack={3}>
+                <CssBaseline/>
                 <App/>
             </SnackbarProvider>
         </QueryClientProvider>
