+++ Client/src/types/credentials.ts
@@ -1,0 +1,4 @@
+export interface Credentials {
+    email: string;
+    password: string;
+}
