--- Server/services/userService/types/user.ts
@@ -1,8 +1,0 @@
-export interface User {
-    firstName: string;
-    lastName: string;
-    email: string;
-    password: string;
-    admin: boolean;
-    _id: string;
-}
